refactor: API呼び出し用ディレクトリを作成
