+++ lib/main_page_layout.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_training/constant/weather_condition.dart';
-import 'package:flutter_training/view_model/view_model.dart';
+import 'package:flutter_training/repository/fetch_yumemi_weather.dart';
 
 /// 大枠のウィジェット
 class MainPageLayout extends StatefulWidget {
